Move flowmodel base visitor and traversing utils to ballerina-core
