+++ workspaces/ballerina/bi-diagram/src/index.tsx
@@ -27,7 +27,7 @@ export { ConnectorIcon } from "./components/ConnectorIcon";
 export type { FlowNodeStyle } from "./utils/types";
 
 // traversing utils
-export { traverseFlow, traverseNode } from "./utils/ast";
+export { traverseFlow, traverseNode } from "@wso2/ballerina-core";
 export { AddNodeVisitor } from "./visitors/AddNodeVisitor";
 export { RemoveNodeVisitor } from "./visitors/RemoveNodeVisitor";
 export { RemoveEmptyNodesVisitor } from "./visitors/RemoveEmptyNodesVisitor";
