+++ workspaces/ballerina/bi-diagram/src/visitors/BreakpointVisitor.ts
@@ -16,8 +16,7 @@
  * under the License.
  */
 
-import { FlowNode, CurrentBreakpointsResponse as BreakpointInfo } from "@wso2/ballerina-core";
-import { BaseVisitor } from "./BaseVisitor";
+import { FlowNode, CurrentBreakpointsResponse as BreakpointInfo, BaseVisitor } from "@wso2/ballerina-core";
 
 
 export class BreakpointVisitor implements BaseVisitor {
